removed the ward entity

+++ src/main/java/com/ke/location/web/rest/dto/CountyDto.java
@@ -2,12 +2,12 @@ package com.ke.location.web.rest.dto;
 
 public interface CountyDto {
     public String getName();
-    public Long getId();
+    public Integer getId();
     public CountyView getCounty();
 
     public interface  CountyView {
 
-        public Long getId();
+        public Integer getId();
 
         public String getName();
     }
